Log batch in read status

+++ TLGX_CONSUMER_SERVICE/DataLayer/DL_UploadStaticData.cs
@@ -466,7 +466,7 @@ namespace DataLayer
                     var skip = RQ.PageSize * RQ.PageNo;
 
                     var AttrMapResult = (from a in AttrMapSearch
-                                         orderby (a.Priority ?? 0) descending, a.AttributeType,a.CREATE_DATE descending //, a.AttributeName, a.AttributeValue
+                                         orderby (a.Priority ?? 0) descending, a.AttributeType, a.CREATE_DATE descending //, a.AttributeName, a.AttributeValue
                                          select new DataContracts.UploadStaticData.DC_SupplierImportAttributeValues
                                          {
                                              SupplierImportAttributeValue_Id = a.SupplierImportAttributeValue_Id,
@@ -759,7 +759,7 @@ namespace DataLayer
                     total = FileSearch.Count();
 
                     var skip = RQ.PageSize * RQ.PageNo;
-                    if(RQ.PageSize==0)
+                    if (RQ.PageSize == 0)
                     {
                         RQ.PageSize = int.MaxValue;
                     }
@@ -1025,7 +1025,7 @@ namespace DataLayer
                 {
                     if (obj.SupplierImportFile_Id != null && !string.IsNullOrWhiteSpace(obj.Step))
                     {
-                        if (obj.CurrentBatch != 0)
+                        if (obj.CurrentBatch != 0 && obj.Step != "READ")
                         {
                             context.SupplierImportFile_Progress.RemoveRange(context.SupplierImportFile_Progress.Where(w => w.SupplierImportFile_Id == obj.SupplierImportFile_Id && w.CurrentBatch != 0 && w.Step.ToString().ToUpper() == obj.Step.ToString().ToUpper() && w.CurrentBatch != obj.CurrentBatch));
                         }
@@ -1038,6 +1038,8 @@ namespace DataLayer
                         {
                             progress.PercentageValue = obj.PercentageValue;
                             progress.LastCheckedOn = DateTime.Now;
+                            progress.CurrentBatch = obj.CurrentBatch;
+                            progress.TotalBatch = obj.TotalBatch;
                         }
                         else
                         {
@@ -2383,7 +2385,7 @@ namespace DataLayer
                 obj.VerboseLog = GetStaticDataUploadVerboseLog(new DataContracts.UploadStaticData.DC_SupplierImportFile_VerboseLog_RQ { SupplierImportFile_Id = fileid });
                 obj.FileStatistics = GetStaticDataUploadStatistics(new DataContracts.UploadStaticData.DC_SupplierImportFile_Statistics_RQ { SupplierImportFile_Id = fileid });
                 obj.FileDetails = GetStaticDataFileDetail(new DataContracts.UploadStaticData.DC_SupplierImportFileDetails_RQ { SupplierImportFile_Id = fileid });
-                obj.ErrorLog= GetStaticDataUploadErrorLog(new DataContracts.UploadStaticData.DC_SupplierImportFile_ErrorLog_RQ {SupplierImportFile_Id=fileid });
+                obj.ErrorLog = GetStaticDataUploadErrorLog(new DataContracts.UploadStaticData.DC_SupplierImportFile_ErrorLog_RQ { SupplierImportFile_Id = fileid });
                 return obj;
             }
             catch (Exception ex)
